requirements.txt
Flask==2.3.3
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3 